requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.6
django-stubs==1.14.0
django-stubs-ext==0.7.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy==0.991
mypy-extensions==1.0.0
python-dotenv==0.21.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.4
typing_extensions==4.4.0
